+++ test/unit/fuzz-tools.test.js
@@ -133,6 +133,7 @@ import * as hemeV132 from '../../lib/heme-v132.js';
 import * as warfarinV133 from '../../lib/warfarin-v133.js';
 import * as oncV134 from '../../lib/onc-v134.js';
 import * as lymphomaV135 from '../../lib/lymphoma-v135.js';
+import * as endoV136 from '../../lib/endo-v136.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -211,6 +212,7 @@ const MODULES = {
   'warfarin-v133.js': warfarinV133,
   'onc-v134.js': oncV134,
   'lymphoma-v135.js': lymphomaV135,
+  'endo-v136.js': endoV136,
   'ems-v149.js': emsV149,
 };